Calculator functionality while using operators continuously
Worked on the actual functionality of the calculator. An object holds
the current number being punched in, always checks for a pressed operator
which will trigger a check if there has been any calculations done and
if so it will accumulate to that value otherwise it will 'initiate' that
value as an integer.