Fixed accumulation bug & added operation check function
When the calculator would be accumulating, using operators in a chained
fashion instead of equals sign, at the last calculation if you pressed equals
it wouldn't calculate. Fixed by adding more logic to the possible states
of the object. Inserted a function to check obvious situations of no math
possibility and reset the calculator.